Fixing bugs in the blocking mapreduce jobs. Introducing V1 of LSHFPS MapReduce

+++ pprl-scaling-framework-mapreduce/pprl-scaling-framework-mapreduce-blocking/src/main/java/gr/upatras/ceid/pprl/mapreduce/CountIdPairsMapper.java
@@ -1,6 +1,6 @@
 package gr.upatras.ceid.pprl.mapreduce;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.ShortWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
@@ -9,9 +9,9 @@ import java.io.IOException;
 /**
  * Count Id pairs mapper.
  */
-public class CountIdPairsMapper extends Mapper <Text,Text,TextPairWritable,IntWritable> {
+public class CountIdPairsMapper extends Mapper <Text,Text,TextPairWritable,ShortWritable> {
 
-    public static IntWritable ONE = new IntWritable(1);
+    public static ShortWritable ONE = new ShortWritable((short)1);
 
     @Override
     protected void map(Text key, Text value, Context context) throws IOException, InterruptedException {
